# docs/backend-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TRUST-TRIAGE Backend API",
    "description": "TRUST-TRIAGE Backend API는 Windows PE(`.exe`, `.dll`)를 접수하고 정적 모델 판정,\n위험 신호, 선택적 심층 분석 증거, 시스템 처리 제안 및 전문가 검토 이력을 제공합니다.\n파일 접수는 비동기 방식이며 `202 Accepted`는 작업 등록을 의미합니다.\n\n### 표준 처리 흐름\n\n| 단계 | API | 계약 |\n| --- | --- | --- |\n| 1. 생존 확인 | [GET /health](./docs#/01.%20%EC%84%9C%EB%B9%84%EC%8A%A4%20%EC%83%81%ED%83%9C/health_health_get) | API 프로세스의 생존 상태를 반환합니다. |\n| 2. 분석 접수 | [POST /analyses](./docs#/02.%20%EB%B6%84%EC%84%9D%20%EC%A0%91%EC%88%98/upload_analyses_post) | PE 파일을 등록하고 `analysis_id`를 반환합니다. |\n| 3. 상태 조회 | [GET /analyses/{analysis_id}/status](./docs#/03.%20%EB%B6%84%EC%84%9D%20%EC%A1%B0%ED%9A%8C/get_status_analyses__analysis_id__status_get) | 비동기 작업의 현재 상태와 단계를 반환합니다. |\n| 4. 결과 조회 | [GET /analyses/{analysis_id}](./docs#/03.%20%EB%B6%84%EC%84%9D%20%EC%A1%B0%ED%9A%8C/get_analysis_analyses__analysis_id__get) | 누적 분석 결과와 최종 처리 제안을 반환합니다. |\n\n### 상태 및 응답 의미\n\n`QUEUED`와 `RUNNING`은 비종료 상태이며, `COMPLETED`와 `FAILED`는 종료 상태입니다.\nHTTP `200`은 조회 요청의 성공을 나타내므로 응답의 `status`가 `FAILED`일 수 있습니다.\nSwagger UI의 `Example Value`는 스키마 예시이며, 실행된 요청의 실제 응답은\n`Server response → Response body`에 표시됩니다. 예시의 식별자, 확률 및 판정값은 실제 분석 결과가 아닙니다.\n\n**인증 정책:** 현재 배포 환경에서는 `X-API-Key` 검증이 비활성화되어 있습니다.\n접수 정보 저장에는 데이터베이스가 필요하고, 비동기 분석 수행에는 별도의 분석 처리기(`run`)가 필요합니다.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/batches": {
      "post": {
        "tags": [
          "04. 배치 분석"
        ],
        "summary": "배치 분석 접수",
        "description": "### 목적\n\n여러 Windows PE 파일을 단일 배치로 등록하고 배치 수준의 추적 식별자를 발급합니다.\n\n### 요청\n\n- Content-Type: `multipart/form-data`\n- 필수 필드: 반복 가능한 `files`\n- 파일 수 상한: **10개**\n- 파일별 크기 상한: **50 MiB (52,428,800 bytes)**\n- `Idempotency-Key`: 동일한 파일 집합과 순서로 구성된 요청의 재전송을 식별하는 선택 헤더\n\n### 응답\n\nHTTP `202 Accepted`는 배치와 파일별 작업이 등록되었음을 의미합니다.\n`batch_id`는 배치 집계 조회에 사용하고, `analyses[].analysis_id`는 개별 파일 조회에 사용합니다.\n\n### 입력별 처리 및 멱등성\n\n지원하지 않는 확장자·잘못된 PE·미지원 PE 형식은 `entries`에 `SKIPPED`와 사유를 기록하고 유효한 PE만 작업으로 등록합니다.\n`input_count = accepted_count + skipped_count`이고 `total_count`는 등록한 분석 수(`accepted_count`)입니다.\n모두 제외된 입력도 배치 접수 내역을 보존하며 `analyses`는 빈 목록입니다. 접수 제한 초과나 저장소·DB 장애는 요청 전체 오류입니다.\n접수 후 개별 분석 실패는 다른 파일의 분석을 중단하지 않습니다.\n동일한 멱등성 키에 다른 파일 집합 또는 순서를 사용하면 `409`가 반환됩니다.",
        "operationId": "upload_batch_batches_post",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "description": "선택적 멱등성 키. 동일한 업로드 요청을 재전송할 때 같은 값을 사용하면 중복 작업 생성을 방지합니다. 동일한 키에 다른 파일 집합을 사용하면 409가 반환됩니다.",
              "title": "Idempotency-Key"
            },
            "description": "선택적 멱등성 키. 동일한 업로드 요청을 재전송할 때 같은 값을 사용하면 중복 작업 생성을 방지합니다. 동일한 키에 다른 파일 집합을 사용하면 409가 반환됩니다."
          }
        ],
        "responses": {
          "202": {
            "description": "비동기 분석 작업이 접수되었습니다. 응답 식별자로 상태와 결과를 조회할 수 있습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchAccepted"
                },
                "examples": {
                  "accepted": {
                    "summary": "묶음 접수 — batch_id와 analysis_id가 함께 반환됨",
                    "value": {
                      "batch_id": "batch_example_001",
                      "total_count": 1,
                      "input_count": 1,
                      "accepted_count": 1,
                      "skipped_count": 0,
                      "archive_file_count": null,
                      "entries": [
                        {
                          "input_index": 0,
                          "filename": "example.exe",
                          "status": "ACCEPTED",
                          "analysis_id": "analysis_example_001",
                          "sha256": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                          "size_bytes": 14848,
                          "reason_code": null,
                          "reason": null
                        }
                      ],
                      "analyses": [
                        {
                          "analysis_id": "analysis_example_001",
                          "batch_id": "batch_example_001",
                          "sha256": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                          "created_at": "2026-09-08T09:00:00Z",
                          "status": "QUEUED",
                          "duplicate_of": null
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류입니다. 장애 분석에는 error.code와 요청 시각이 필요합니다.",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INTERNAL_ERROR",
                    "message": "서버 처리 중 오류가 발생했습니다.",
                    "retryable": false
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "요청 본문을 해석할 수 없습니다. 경로, HTTP 메서드 또는 multipart 형식이 유효하지 않습니다.",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_HTTP_REQUEST",
                    "message": "요청 경로, 메서드 또는 multipart 형식이 유효하지 않습니다.",
                    "stage": null,
                    "retryable": false
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증에 실패했습니다. 유효한 X-API-Key가 필요합니다.",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "유효한 X-API-Key가 필요합니다.",
                    "stage": null,
                    "retryable": false
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "408": {
            "description": "설정된 제한 시간 안에 파일 전송이 완료되지 않았습니다.",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "UPLOAD_TIMEOUT",
                    "message": "요청 본문 업로드 시간이 초과되었습니다.",
                    "stage": "UPLOAD",
                    "retryable": false
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "현재 리소스 상태 또는 멱등성 키에 연결된 기존 요청과 충돌합니다. error.code가 충돌 유형을 구분합니다.",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "IDEMPOTENCY_CONFLICT",
                    "message": "같은 접수 키에 다른 파일이 연결되어 있습니다.",
                    "stage": null,
                    "retryable": false
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "파일 또는 전체 요청 본문이 서버의 업로드 제한을 초과했습니다.",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "FILE_TOO_LARGE",
                    "message": "파일이 허용 크기를 초과했습니다.",
                    "stage": "UPLOAD",
                    "retryable": false
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "요청 값이 API 계약을 충족하지 않습니다. 필수 필드, 식별자 형식, 파일 형식 또는 숫자 범위가 유효하지 않습니다.",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_REQUEST",
                    "message": "요청 필드 형식 또는 필수 값이 유효하지 않습니다.",
                    "stage": null,
                    "retryable": false
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "서비스 의존성이 준비되지 않았습니다. error.code가 데이터베이스, 저장소 등 실패한 의존성을 식별합니다.",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "DATABASE_NOT_CONFIGURED",
                    "message": "BACKEND_DATABASE_URL이 설정되지 않았습니다.",
                    "stage": null,
                    "retryable": false
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "description": "multipart/form-data 형식의 PE 파일 본문입니다. 바이너리를 JSON으로 인코딩하지 않습니다.",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "files"
                ],
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary",
                      "description": "분석 대상 Windows PE 파일입니다. 지원 확장자는 .exe와 .dll입니다."
                    },
                    "description": "배치 분석 대상 파일 목록입니다. 파일 수와 파일별 용량은 서버 설정의 상한을 적용합니다."
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchAccepted": {
        "properties": {
          "input_count": {
            "type": "integer",
            "title": "Input Count",
            "description": "접수 전에 검사한 파일 수입니다. 폴더는 제외합니다.",
            "default": 0
          },
          "accepted_count": {
            "type": "integer",
            "title": "Accepted Count",
            "description": "독립 분석 작업을 등록한 파일 수입니다.",
            "default": 0
          },
          "skipped_count": {
            "type": "integer",
            "title": "Skipped Count",
            "description": "분석 작업을 만들지 않은 파일 수입니다.",
            "default": 0
          },
          "archive_file_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archive File Count",
            "description": "ZIP 내부 파일 수. 일반 다중 입력이면 null입니다."
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/BatchInputEntry"
            },
            "type": "array",
            "title": "Entries"
          },
          "batch_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,128}$",
            "title": "Batch Id",
            "description": "묶음 전체를 조회할 때 사용하는 번호입니다."
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count",
            "description": "이번 묶음으로 등록한 분석 수입니다. accepted_count와 같으며 SKIPPED는 제외합니다."
          },
          "analyses": {
            "items": {
              "$ref": "#/components/schemas/AnalysisAccepted"
            },
            "type": "array",
            "title": "Analyses",
            "description": "파일별 접수 결과입니다. 각 항목에 고유한 analysis_id가 있습니다."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "batch_id",
          "total_count",
          "analyses"
        ],
        "title": "BatchAccepted"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "오류 유형을 식별하는 안정적인 코드입니다. 로그 상관관계 분석과 장애 지원에 사용합니다."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "무엇이 잘못됐는지 설명하는 메시지입니다."
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage",
            "description": "오류가 발생한 단계입니다. 특정 단계와 관련 없으면 null입니다."
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "description": "서버가 일시적 오류로 분류했는지 나타냅니다. true여도 즉시 성공을 보장하지 않습니다.",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "description": "API 접근 인증에 사용하는 X-API-Key 헤더입니다. 서버의 API 키 검증이 비활성화된 환경에서는 생략할 수 있습니다.",
        "in": "header",
        "name": "X-API-Key"
      },
      "ReviewerKeyAuth": {
        "type": "apiKey",
        "description": "전문가 검토 등록 권한을 검증하는 X-Reviewer-Key 헤더입니다. API 키 검증도 활성화된 환경에서는 두 인증 헤더가 모두 필요합니다.",
        "in": "header",
        "name": "X-Reviewer-Key"
      }
    }
  },
...
}